pyproject.toml
[project]
name = "ia-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.120.4",
    "langchain>=1.0.3",
    "langchain-core>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langgraph-cli[inmem]>=0.4.4",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]

[tool.black]
line-length = 100
target-version = ['py312']
skip-string-normalization = true


